feat - add github actions
