+++ .github/workflows/pr-qual.yml
@@ -1,0 +1,79 @@
+name: Flow to merge to quality from develop
+
+on:
+  pull_request:
+    branches:
+      - qual
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Cache Maven packages
+        uses: actions/cache@v2
+        with:
+          path: |
+            ~/.m2
+          key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
+          restore-keys: |
+            ${{ runner.os }}-m2
+      - name: Install Protocol Buffers Compiler
+        run: sudo apt-get update && sudo apt-get install -y protobuf-compiler
+      - name: Check protoc version
+        run: protoc --version
+      - name: Set up JDK 17
+        uses: actions/setup-java@v2
+        with:
+          java-version: '17'
+          distribution: 'temurin'
+      - name: Build with Maven
+        run: mvn -B package -Dmaven.test.skip=true
+  unit-test:
+    needs: build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Cache Maven packages
+        uses: actions/cache@v2
+        with:
+          path: |
+            ~/.m2
+          key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
+          restore-keys: |
+            ${{ runner.os }}-m2
+      - name: Install Protocol Buffers Compiler
+        run: sudo apt-get update && sudo apt-get install -y protobuf-compiler
+      - name: Check protoc version
+        run: protoc --version
+      - name: Set up JDK 17
+        uses: actions/setup-java@v2
+        with:
+          java-version: '17'
+          distribution: 'temurin'
+      - name: Test with Maven
+        run: mvn test -Punit-tests -Dopenapi.generator.skip=true
+  integration-test:
+    needs: build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Cache Maven packages
+        uses: actions/cache@v2
+        with:
+          path: |
+            ~/.m2
+          key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
+          restore-keys: |
+            ${{ runner.os }}-m2
+      - name: Install Protocol Buffers Compiler
+        run: sudo apt-get update && sudo apt-get install -y protobuf-compiler
+      - name: Check protoc version
+        run: protoc --version
+      - name: Set up JDK 17
+        uses: actions/setup-java@v2
+        with:
+          java-version: '17'
+          distribution: 'temurin'
+      - name: Test with Maven
+        run: mvn verify -Pintegration-tests -Dopenapi.generator.skip=true
