+++ .github/workflows/push-dev.yml
@@ -1,0 +1,17 @@
+name: Auto PR to quality from development when merge
+
+on:
+  push:
+    branches:
+      - develop
+
+jobs:
+  release:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Create PR
+        uses: poorva17/create-pr-action@main
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          HEAD_BRANCH: develop
+          BASE_BRANCH: qual
